move TransformerManager to transformations module
