+++ bids/analysis/__init__.py
@@ -1,4 +1,9 @@
 from .analysis import Analysis
 from .auto_model import auto_model
+from .transformations import TransformerManager
 
-__all__ = ['Analysis', 'auto_model']
+__all__ = [
+    'Analysis',
+    'auto_model',
+    'TransformerManager'
+]
